requirements.txt
Django
requests
groq
python-dotenv
pprintpp